Correct type

+++ tests/internal/scripts/mpi/bcast.py
@@ -29,7 +29,7 @@ if __name__ == '__main__':
     if rank == master:
         msg = rank + 1000
     else:
-        msg = 0
+        msg = np.int32(0)
 
     length = np.int32(1)
     mpi_bcast (msg, length, MPI_INTEGER8, master, comm, ierr)
